requirements.txt
cloudscraper
python-dotenv